refactor: 코드래빗 의견 반영

+++ src/app/badge/util.tsx
@@ -15,7 +15,7 @@ export const fontStyle = <T extends keyof typeof FONTS>(
 
   return {
     style: { fontFamily: `Noto${fontWeight}` },
-    tw: `text-[${parseInt(size)}px] leading-[${parseInt(lineHeight)}px] text-[${color}] ${tw}`,
+    tw: `text-[${parseInt(size)}px] leading-[${parseFloat(lineHeight)}px] text-[${color}] ${tw}`,
   };
 };
